+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   Markdown.  Code output in raw format (interpreted as Markdown) is no longer
   lost when converting to document formats other than Markdown (#26).
 * Added support for SageMath (#5).
+* Added support for remaining unsupported Pandoc command-line options,
+  including `--defaults` (#14).
 * Julia now uses `--project=@.` (#10).
 * Documentation now includes details of code execution and how this can result
   in different output compared to interactive sessions (#11).